- Refactored StringHelper to GenresManager, so it handles the responsibility of genres to fetch and convert data, so not just upcoming movies can access it. - Removed responsibility of genres from upcomingviewmodel - Landscape view of collectionview in UpcomingViewModelController - Refactored Details controller - Readme updated
